requirements.txt
aiohttp==3.8.1
python-dotenv==0.19.2
openai==0.27.0
python-slugify==5.0.2
beautifulsoup4==4.10.0
requests==2.26.0